refactor: common 패키지 추가
