+++ src/main/java/prezwiz/server/common/annotation/ExeTimer.java
@@ -1,4 +1,4 @@
-package prezwiz.server.annotation;
+package prezwiz.server.common.annotation;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
