+++ src/main/java/prezwiz/server/common/exception/MemberNotFoundException.java
@@ -1,4 +1,4 @@
-package prezwiz.server.exception;
+package prezwiz.server.common.exception;
 
 public class MemberNotFoundException extends RuntimeException {
 }
