+++ src/main/java/prezwiz/server/common/util/LocalPrezStorage.java
@@ -1,4 +1,4 @@
-package prezwiz.server.util;
+package prezwiz.server.common.util;
 
 import lombok.extern.slf4j.Slf4j;
 import org.apache.poi.xslf.usermodel.XMLSlideShow;
