+++ src/main/java/prezwiz/server/service/prez/GPTPrezService.java
@@ -14,9 +14,8 @@ import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.util.StopWatch;
 import org.springframework.web.reactive.function.client.WebClient;
-import prezwiz.server.annotation.ExeTimer;
+import prezwiz.server.common.annotation.ExeTimer;
 import prezwiz.server.dto.request.CreateRequestDto;
 import prezwiz.server.dto.response.PresentationResponseDto;
 import prezwiz.server.dto.slide.SlideDto;
@@ -30,8 +29,8 @@ import prezwiz.server.entity.Member;
 import prezwiz.server.entity.Presentation;
 import prezwiz.server.repository.MemberRepository;
 import prezwiz.server.repository.PresentationRepository;
-import prezwiz.server.util.PrezStorage;
-import prezwiz.server.util.SlideUtil;
+import prezwiz.server.common.util.PrezStorage;
+import prezwiz.server.common.util.SlideUtil;
 import reactor.core.publisher.Mono;
 
 import java.io.ByteArrayOutputStream;
